requirements.txt
# --- Core Orchestration & AI Framework ---
langchain
langchain-core
langchain-community
langgraph
mistralai
langchain-mistralai
google-genai

# --- Vector & NLP (BGE-M3 Multilingual) ---
sentence-transformers
FlagEmbedding
torch
numpy

# --- Data Tiers (SQL & Graph via Supabase) ---
supabase
psycopg2-binary
upstash-redis

# --- Backend & API ---
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv

# --- Evaluation & Analysis ---
ragas
pandas
datasets

# --- Utilities & Data Generation ---
faker
uuid